+++ Apps-Script/server/test/helpers/string-utils.test.js
@@ -114,7 +114,7 @@ const StringUtilsTest = (() => {
 
 if (typeof module !== "undefined") {
   UnitTestingApp = require('../unit-testing-app.js');
-  MockData = require('../mock-data.js');
+  MockData = require('../../lib/mock-data.js');
   StringUtils = require('../../helpers/string-utils.js');
   module.exports = StringUtilsTest;
 }
